REFAC: postResult thunk added to results.js, modified fetchResults thunk
