+++ store/time.js
@@ -1,31 +1,13 @@
-import axios from 'axios';
-import { fetchResultsThunk } from './results';
-import domain from '../domain.js';
-
 // ACTION TYPES
 const INCREMENT_TIME = 'INCREMENT_TIME';
 const RESET_TIME = 'RESET_TIME';
-
-// INITIAL STATE
-const defaultTime = 0;
 
 // ACTION CREATORS
 export const incrementTime = time => ({ type: INCREMENT_TIME, time });
 export const resetTime = () => ({ type: RESET_TIME });
 
-// THUNK CREATORS
-export const postResult = (name, time, userId, mapId) => dispatch => {
-  return axios
-    .post(`${domain}/api/results`, { name, time, userId, mapId })
-    .then(() => {
-      dispatch(resetTime());
-      dispatch(fetchResultsThunk());
-    })
-    .catch(err => console.error(err));
-};
-
 // REDUCER
-export default function(state = defaultTime, action) {
+export default function(state = 0, action) {
   switch (action.type) {
     case INCREMENT_TIME:
       return action.time;
